requirements.txt
python-telegram-bot==20.6
rembg==2.0.50
